shared internal buffer.
When a template is called, if vash.helpers.__vo exists, it will be
used as the output buffer. This allows for nested template calls to have
basic awareness and composability. For example, a view engine could use
this to call a layout template from within an inner template, and the
layout could then wrap around the inner template with some work.

More on this shortly.