Jade-like extends/block/append/prepend/include
vash.helpers.layout provides very primitive view engine support for
Express 3.x. There is a small demo located at `test/vash.express.js`.

There are a few limitations, specifically around blocks. In Jade, the
"highest" block or the most "parent" block can have default content.
Vash's implementation happens at runtime, not at parse/compile time.
Because of this, the root block definition must not have default
content, so that Vash knows that _this_ is where the output should be
placed.

The only other limitation currently is that all content function blocks
(see the demo for examples) must have `modelName` (which is by default
`model`) defined as a parameter. Again, because this is all happening at
runtime, it was necessary to preserve scoping without resorting to
`with` or other ugly-ish hacks.

Browser support has not been tested fully, so while it theoretically
*should* work even in a browser (assuming all templates are loaded), it
is not guaranteed.