@: paired with \n. tok.val defaults to ''
Defaulting tok.val is to prevent tokens that are added to the AST, such
as AT_COLON, but whose lexer token definition has no capture groups,
from throwing an error when .toString() is called. AT_COLON is basically
the only token that appears in the AST but has no output value.