@: opens a MKP block, closed by \n (NEWLINE)
This preserves proper behavior, but also means that a crazy template
like

	`@if(true){ @:This is text }`

will now fail (as it should) due to a missing newline character after
the final content. A proper version:

	`@if(true){ @:This is text\n }`

I always try to preserve the ability for Vash to be whitespace
insignificant, but I made an exception in this case because the @: is
inherently whitespace (via newline) significant. It creates a block
where there are no triggers to do so.