compiler only does dev replace for non-tpl code
An executing template now returns `vash.Helpers` instance, allowing for
more complex things in the future, like implementing promises.
`toString` and `toHtmlString` are implemented to allow coercion to
convert the instance to the template output string.

`Buffer.toString` does not empty the buffer, because should reading a
value destroy it? Luckily this isn't quantum mechanics.