layout helpers render immediately, not deferred. #16
There may still be issues with `include` in the event that it calls
`extend` from within, but that behavior is basically undefined.

`include` should _probably_ work differently. They could be in their own
context, and not share blocks, OR they could accept a callback block.
This would allow for appending to the included content, and even
overriding blocks only defined in the included template.