`vash.batch`, `vash.[un]install`, `vash.lookup`
`vash.batch` allows several templates to be defined in a "single" call
to `vash.compile`, and have them automatically be "installed" into the
template cache, to be later retrieved by `vash.lookup`.

In order to support this reconstitution, the `modelName` and
`helpersName` configuration values must be passed into `vash.link`,
which is now defined in the runtime, but is also used by the compiler.

`tpl.toClientString` will still handle all of this automatically.