`vash.config.simple` optimizes for fast path
Default is false. Setting to true bypasses the following:

- `onRenderEnd` callbacks are completely ignored
- The helpers instance becomes a plain object with a `buffer` property
  set to a plain array, and `raw` and `escape` references. Thus the
  entire buffer api is ignored, and ALL helpers other than `raw` and
  `escape` are not available.
- `asContext` render option is completely ignored.
- `vash.config.htmlEscape` and `vash.config.useWith` still behave as
  expected.

This option is available to provide the fastest possible template
rendering for those users who do not need the amenities of the Helpers
context and Buffer api, and would like to sacrifice them for speed.

This is a compile-time option, and thus a template compiled with this
option cannot be converted to a "full" template.