`vash.compileHelper` and `vash.batch` are at compile time
New batch syntax:
-----------------

	// @batch = path/to/tpl
	<p>@model.what</p>

	// @batch = path/to/another/tpl
	<p>@model.another</p>

Only rules are that it must be something like, whitespace optional:

	//@batch=whatever

This syntax was chosen because it's relatively simple, and also does not
require complex parsing; it's easy to split the templates by a simple
regex, and then feed them individually into `vash.compile`. The above
example would return an object that looks like:

	{
		'path/to/tpl': function(model, options){ ... },
		'path/to/another/tpl': function(model, options){ ... }
	}

Compiled Helpers
----------------

Helpers are most easily created via a simple assignment to a property of
`vash.helpers`. But if a helper requires complex markup, then the user
is back to string concatenation, which is no fun. Instead, compiled
helpers let you use the power of vash's syntax!

Here's an example of a footnote helper:

	vash.helpers.footnote = function(id, ctn){
		this.fnCounter = this.fnCounter || 0;
		this.fnIds = this.fnIds || {};
		this.fnCtn = this.fnCtn || {};

		if(ctn){
			<li id="fn:@id">
				@ctn()
				<a rev="footnote" href="#fnref:@id">↩</a>
			</li>
		} else {
			this.fnIds[id] = ++this.fnCounter;
			<sup id="fnref:@id">
				<a rel="footnote" href="#fn:@id">@html.raw(this.fnCounter)</a>
			</sup>
		}
	}

It can be called in one of two ways:

	// add a [1] to the end of a sentence or something
	@html.footnote('note about somebody');

	// define the content of the footnote. it can contain any content you want!
	@html.footnote('note about somebody', function(){
		<text>
			This is the footnote about that person named "somebody", which is
			a bit silly and altogether extremely confusing.
		</text>
	})

To compile this helper, you simply call `vash.compileHelper`, and it will
automatically be attached to vash.helpers. There are a few limitations due
to how the simple string manipulation works. Those are:

	* The definition must be a simple function assignment, and whitespace
	is optional:
		vash.helpers.NAME=function OPTNAME(ARGS...){}
	* More complex expressions, like:
		vash.helpers.what = (function(){ return function(){ .. } }())
	will fail hard. If you need this functionality, fallback to hand-
	defining a helper using the Buffer API.