a self-closing tag should not always exit curr ast
If the parent is BLK, a self closing tag should only exit if the current
node does not have a tag name. This prevents the self-closing tag from
closing the current markup block, switching to BLK, and then switching
back to the closing tag. It also prevents a self-closing tag closing the
curent markup block, and then subsequent markup content getting treated
as BLK:

	@if(true){ <a>Hello <br /> world!</a> }

Was throwing an error, trying to use `world` as BLK.