vash(1) --helpers <file>
A file can be passed in, and thus when paired with --render, can enable
helpers via command line. A complex example:

	cat myTpl.vash | bin/vash --render --helpers \
		<(cat helpers/* | bin/vash --helper)

This example grabs all the files in the helpers directory, compiles them
as helpers, and then passes them as a temporary named pipe into another
call to vash(1) that is rendering myTpl.vash.