Unified non-void and void html tags (self-closing)
A void html element, such as <br />, previously could not contain
something like <br @(what ? sure : ok) />. This was because the logic
for self-closing tags was separate from the much more complex
HTML_TAG_OPEN lexing. Now an HTML_TAG_VOID_OPEN is checked for first,
and if found, the ast is marked via `tagVoid`.