For loop prevents max args limit. Fixes #29
JS engines, _of course_ have a maximum argument limit for function
calls. This is undocumented, but differs per platform and depends on
memory availability and engine tuning. On real-world templates, the use
of apply in debug mode was causing large arrays of strings to be passed
around as arguments, causing max call stack exceptions (v8) and likely a
specific too many args error in FF.